+++ src/configs/networks/types/ETH.js
@@ -1,3 +1,6 @@
+import tokens from '@/tokens/tokens-eth.json'
+import contracts from '@/contracts/contract-abi-eth.json'
+
 export default {
   name: 'ETH',
   name_long: 'Ethereum',
@@ -5,6 +8,6 @@ export default {
   blockExplorerTX: 'https://etherscan.io/tx/[[txHash]]',
   blockExplorerAddr: 'https://etherscan.io/address/[[address]]',
   chainID: 1,
-  tokens: [],
-  contracts: []
+  tokens: tokens,
+  contracts: contracts
 }