+++ src/configs/networks/types/ROP.js
@@ -1,10 +1,13 @@
+import tokens from '@/tokens/tokens-rop.json'
+import contracts from '@/contracts/contract-abi-rop.json'
+
 export default {
-  name: 'Ropsten',
+  name: 'ROP',
   name_long: 'Ropsten',
   homePage: 'https://github.com/ethereum/ropsten',
   blockExplorerTX: 'https://ropsten.etherscan.io/tx/[[txHash]]',
   blockExplorerAddr: 'https://ropsten.etherscan.io/address/[[address]]',
   chainID: 3,
-  tokens: [],
-  contracts: []
+  tokens: tokens,
+  contracts: contracts
 }