+++ src/configs/networks/types/index.js
@@ -1,6 +1,38 @@
+import AKA from './AKA.js'
+import CLO from './CLO.js'
+import EGEM from './EGEM.js'
+import ELLA from './ELLA.js'
+import EOSC from './EOSC.js'
+import ESN from './ESN.js'
+import ETC from './ETC.js'
 import ETH from './ETH.js'
-import Ropsten from './Ropsten.js'
+import ETSC from './ETSC.js'
+import EXP from './EXP.js'
+import GO from './GO.js'
+import KOV from './KOV.js'
+import MUSIC from './MUSIC.js'
+import POA from './POA.js'
+import RIN from './RIN.js'
+import ROP from './ROP.js'
+import TOMO from './TOMO.js'
+import UBQ from './UBQ.js'
 export {
+  AKA,
+  CLO,
+  EGEM,
+  ELLA,
+  EOSC,
+  ESN,
+  ETC,
   ETH,
-  Ropsten
+  ETSC,
+  EXP,
+  GO,
+  KOV,
+  MUSIC,
+  POA,
+  RIN,
+  ROP,
+  TOMO,
+  UBQ
 }