+++ src/main.js
@@ -12,13 +12,17 @@ import store from './store'
 import VueI18n from 'vue-i18n'
 import Vuex from 'vuex'
 import VueQrcode from '@xkeshi/vue-qrcode'
+import PopOver from '@/components/PopOver'
 
 // etc
 import languages from './translations'
 
+Vue.prototype.$eventHub = new Vue()
+
 // Regular Components
 Vue.component('infinite-slider', InfiniteSlider)
 Vue.component(VueQrcode.name, VueQrcode)
+Vue.component('popover', PopOver)
 
 Vue.config.productionTip = false