+++ CHANGELOG.md
@@ -1,4 +1,18 @@
-### Release v0.0.2
 ### Devop
+- Fetch tokens and contract abi [#41](https://github.com/MyEtherWallet/MyEtherWallet/pull/41)
+### Feature
+- Intercept and override signTransaction and signMessage to force confirm modal display and approval
+- Integrate Ledger hardware wallet
+
+### Release v0.0.2
+### Feature
+- additional pages [#28](https://github.com/MyEtherWallet/MyEtherWallet/pull/28)
+- Adding custom tokens and Adding custom networks [#27](https://github.com/MyEtherWallet/MyEtherWallet/pull/27)
+
+### Devop
+- Update footer to match design [#29](https://github.com/MyEtherWallet/MyEtherWallet/pull/29)
 - Miscellaneous cleanup and changes. [#22](https://github.com/MyEtherWallet/MyEtherWallet/pull/22)
+
+### Release v0.0.1
+### Devop
 - Miscellaneous cleanup and changes. [#14](https://github.com/MyEtherWallet/MyEtherWallet/pull/14)